requirements.txt
asgiref==3.5.0
click==8.0.3
cmake==3.22.1
colorama==0.4.4
Django==4.0.1
djangorestframework==3.13.1
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
